+++ webapp/create-test-reports.sh
@@ -7,12 +7,12 @@ pushd ..
     pnpm --filter \!@eighty4/plunder-webapp build 
 popd
 
-node ../cli/lib/plunder.js \
+node ../cli/lib_js/plunder.js \
     --css-breakpoints \
     -o .playwright/plunder/single-page \
     https://alistapart.com
 
-node ../cli/lib/plunder.js \
+node ../cli/lib_js/plunder.js \
     -d "iphone 15 pro max" \
     -o .playwright/plunder/multiple-pages \
     https://alistapart.com \